Function ad-hoc serialization.

+++ kernel/symbol.h
@@ -58,7 +58,7 @@ namespace vita
 
     /// The arity of a function is the number of inputs to or arguments of that
     /// funtion.
-    virtual unsigned arity() const = 0;
+    virtual size_t arity() const = 0;
 
     /// Calculates the value of / performs the action associated to the symbol
     /// (it is implementation specific).
